feat: Add "clear-value" prop for clearable behavior #1157
