+++ src/mixins/input.js
@@ -1,3 +1,5 @@
+import { stopAndPrevent } from '../../utils/event'
+
 export default {
   props: {
     autofocus: [Boolean, String],
@@ -5,7 +7,8 @@ export default {
     maxLength: [Number, String],
     maxHeight: Number,
     placeholder: String,
-    loading: Boolean
+    loading: Boolean,
+    clearValue: {}
   },
   data () {
     return {
@@ -32,6 +35,13 @@ export default {
     },
     select () {
       this.$refs.input.select()
+    },
+    clear (evt) {
+      stopAndPrevent(evt)
+      const val = this.clearValue || null
+      this.$emit('input', val)
+      this.$emit('change', val)
+      this.$emit('clear')
     },
 
     __onFocus (e) {
